requirements.txt
python-telegram-bot==22.7
httpx==0.27.2
python-dotenv==1.0.1
pydub==0.25.1
openai-whisper==20231117
langgraph==0.3.21
aiosqlite==0.21.0
langgraph-checkpoint-sqlite==2.0.10
